Removed non-wizard references to reverse engineering
