+++ src/GUI/EFCorePowerTools/EFCorePowerTools.csproj
@@ -57,6 +57,15 @@
     <Compile Include="..\..\Nupkg\ErikEJ.EntityFrameworkCore.DgmlBuilder\DgmlBuilder.cs">
       <Link>DgmlBuilder\DgmlBuilder.cs</Link>
     </Compile>
+    <Compile Include="Dialogs\EfCoreModelDialog.xaml.cs">
+      <DependentUpon>EfCoreModelDialog.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Dialogs\PickServerDatabaseDialog.xaml.cs">
+      <DependentUpon>PickServerDatabaseDialog.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Dialogs\PickTablesDialog.xaml.cs">
+      <DependentUpon>PickTablesDialog.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Westwind.WPF.Statusbar\StatusBarControl.xaml.cs">
       <DependentUpon>StatusBarControl.xaml</DependentUpon>
     </Compile>
@@ -90,9 +99,6 @@
     <Compile Include="Dialogs\CompareResultDialog.xaml.cs">
       <DependentUpon>CompareResultDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Dialogs\EfCoreModelDialog.xaml.cs">
-      <DependentUpon>EfCoreModelDialog.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Dialogs\PickProjectDialog.xaml.cs">
       <DependentUpon>PickProjectDialog.xaml</DependentUpon>
     </Compile>
@@ -102,14 +108,8 @@
     <Compile Include="Dialogs\PickConfigDialog.xaml.cs">
       <DependentUpon>PickConfigDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Dialogs\PickServerDatabaseDialog.xaml.cs">
-      <DependentUpon>PickServerDatabaseDialog.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Dialogs\CompareOptionsDialog.xaml.cs">
       <DependentUpon>CompareOptionsDialog.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Dialogs\PickTablesDialog.xaml.cs">
-      <DependentUpon>PickTablesDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="EFCorePowerToolsPackage.cs" />
     <Compile Include="GlobalSuppressions.cs" />
@@ -564,19 +564,19 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Dialogs\PickServerDatabaseDialog.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-    </Page>
     <Page Include="Dialogs\CompareOptionsDialog.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Dialogs\PickTablesDialog.xaml">
+    <Page Include="Dialogs\Converter.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Dialogs\Converter.xaml">
+    <Page Include="Dialogs\PickServerDatabaseDialog.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Dialogs\PickTablesDialog.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
