+++ src/GUI/Shared/Shared.projitems
@@ -82,7 +82,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\DacpacAnalyzerHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\ReverseEngineerUserOptions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\DatabaseDiagramHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\RevEngWizardHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\ReverseEngineerHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\SqlConnectionStringBuilderHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\SqlProjHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\FocusExtensions.cs" />
@@ -105,7 +105,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\NupkgExtractor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\ResultDeserializer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\ReverseEngineerCommandOptionsExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\ReverseEngineerHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Handlers\ReverseEngineer\TableListBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\CredentialStore.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\VSHelper.cs" />
